requirements.txt
flask==1.1.2
flask-restful==0.3.8
pymongo==3.11.2
dnspython==2.0.0
bcrypt==3.2.0